feat: rework message generation
Okay, I think the gpt4free repo is back, I thought it died lol. So they've added a whole package, called the same name - gpt4free, and just by installing it literally with one line of code you can generate a message with just a provider name. And I found that You provider works the best, however, when it tries to answer not in English, it just gives some giberish, so to combat that I integrated the translation feature. Unfortunately not via DeepL, because I think the API is paid, but via Google Translate, which is good enough. I also detect input text right inside of the function to check if its in Ukrainian. If it is, we translate it to English and give it as prompt. If not, we just give the original message as prompt. Then, after it generated, we check if the response is in English, which it always is, now that I'm thinking about it, so we can not even check it, and if the input prompt language is Ukrainian, we translate the output to Ukrainian and return it. The next commit will probably be just refactoring that same function, I think I know how to reduce the code amount for it